feat(crypto): add HMAC builtin and modularize builtins into directory
Add crypto.hmac(key, data, alg?, raw?) supporting all 8 hash algorithms with binary-safe key/data handling for AWS Sig V4 key chaining. Manual RFC 2104 HMAC using existing sha2/sha3 crates — zero new dependencies.

Split monolithic builtins.rs (1788 lines) into src/lua/builtins/ directory with 10 focused modules: http, json, serialization, assert, crypto, db, ws, template, core, mod. Pure refactoring — no behavior change.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)
