+++ tests/crypto.rs
@@ -208,3 +208,71 @@ async fn test_invalid_length() {
     assert!(run_lua(r#"crypto.random(0)"#).await.is_err());
     assert!(run_lua(r#"crypto.random(-1)"#).await.is_err());
 }
+
+#[tokio::test]
+async fn test_hmac_sha256_basic() {
+    let result: String = eval_lua(r#"return crypto.hmac("Jefe", "what do ya want for nothing?", "sha256")"#).await;
+    assert_eq!(
+        result,
+        "5bdcc146bf60754e6a042426089575c75a003f089d2739839dec58b964ec3843"
+    );
+}
+
+#[tokio::test]
+async fn test_hmac_sha256_raw_output() {
+    let vm = create_vm();
+    let raw_bytes: mlua::String = vm
+        .load(r#"return crypto.hmac("Jefe", "what do ya want for nothing?", "sha256", true)"#)
+        .eval_async()
+        .await
+        .unwrap();
+    assert_eq!(raw_bytes.as_bytes().len(), 32);
+}
+
+#[tokio::test]
+async fn test_hmac_sha256_key_chaining() {
+    let script = r#"
+        local k1 = crypto.hmac("AWS4" .. "secret", "20130524", "sha256", true)
+        local k2 = crypto.hmac(k1, "us-east-1", "sha256", true)
+        local k3 = crypto.hmac(k2, "s3", "sha256", true)
+        local k4 = crypto.hmac(k3, "aws4_request", "sha256", true)
+        return #k4
+    "#;
+    let len: i64 = eval_lua(script).await;
+    assert_eq!(len, 32);
+}
+
+#[tokio::test]
+async fn test_hmac_default_algorithm() {
+    let result: String = eval_lua(r#"return crypto.hmac("key", "data")"#).await;
+    let explicit: String = eval_lua(r#"return crypto.hmac("key", "data", "sha256")"#).await;
+    assert_eq!(result, explicit);
+}
+
+#[tokio::test]
+async fn test_hmac_sha512() {
+    let result: String =
+        eval_lua(r#"return crypto.hmac("Jefe", "what do ya want for nothing?", "sha512")"#).await;
+    assert_eq!(result.len(), 128);
+    assert_eq!(
+        result,
+        "164b7a7bfcf819e2e395fbe73b56e0a387bd64222e831fd610270cd7ea2505549758bf75c05a994a6d034f65f8f0e6fdcaeab1a34d4a6b4b636e070a38bce737"
+    );
+}
+
+#[tokio::test]
+async fn test_hmac_unsupported_algorithm() {
+    let result = run_lua(r#"crypto.hmac("key", "data", "md5")"#).await;
+    assert!(result.is_err());
+    let err = result.unwrap_err().to_string();
+    assert!(err.contains("unsupported algorithm"), "got: {err}");
+}
+
+#[tokio::test]
+async fn test_hmac_empty_data() {
+    let result: String = eval_lua(r#"return crypto.hmac("key", "")"#).await;
+    assert_eq!(
+        result,
+        "5d5d139563c95b5967b9bd9a8c9b233a9dedb45072794cd232dc1b74832607d0"
+    );
+}
